third_party/blink/web_tests/external/wpt/css/css-anchor-position/position-area-anchor-outside-expected.txt

This is a testharness.js-based test.
[FAIL] Offsets for position-area: span-all
  assert_equals: Check expected offsetLeft expected -200 but got 0
[FAIL] Offsets for position-area: left span-all
  assert_equals: Check expected offsetHeight expected 600 but got 400
[FAIL] Offsets for position-area: span-left span-all
  assert_equals: Check expected offsetLeft expected -200 but got -100
[FAIL] Offsets for position-area: span-all center
  assert_equals: Check expected offsetHeight expected 600 but got 400
[FAIL] Offsets for position-area: span-right span-all
  assert_equals: Check expected offsetHeight expected 600 but got 400
[FAIL] Offsets for position-area: right span-all
  assert_equals: Check expected offsetHeight expected 600 but got 400
[FAIL] Offsets for position-area: top span-all
  assert_equals: Check expected offsetLeft expected -200 but got 0
[FAIL] Offsets for position-area: span-top span-all
  assert_equals: Check expected offsetLeft expected -200 but got 0
[FAIL] Offsets for position-area: center span-all
  assert_equals: Check expected offsetLeft expected -200 but got 0
[FAIL] Offsets for position-area: span-bottom span-all
  assert_equals: Check expected offsetLeft expected -200 but got 0
[FAIL] Offsets for position-area: bottom span-all
  assert_equals: Check expected offsetLeft expected -200 but got 0
Harness: the test ran to completion.
